% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install project requirements}
\usage{
install(cran, bioc, github, lib_cran, lib_bioc, lib_github,
  check_only = FALSE)
}
\arguments{
\item{cran}{\code{[character]} packages to install from a CRAN like repository.}

\item{bioc}{\code{[character]} packages to install from Bioconductor.}

\item{github}{\code{[character]} packages to install from Github.}

\item{lib_cran}{\code{[character(1)]} where to install packages from CRAN like repositories.}

\item{lib_bioc}{\code{[character(1)]} where to install packages from Bioconductor.}

\item{lib_github}{\code{[character(1)]} where to install packages from GitHub
(their dependencies will also be installed here).}

\item{check_only}{\code{[logical(1)]} should installation be skipped?}
}
\value{
Returns a \code{[character]} vector of packages that need/needed to be
installed. These packages will be installed if \code{check_only} is \code{FALSE}
(the default).
}
\description{
Install project requirements
}
