% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_from_github}
\alias{install_from_github}
\title{Install from GitHub style repositories}
\usage{
install_from_github(urls, lib, check_only = TRUE)
}
\arguments{
\item{urls}{\code{[character]} GitHub repository URLs (e.g. username/reponame@commit)}

\item{lib}{\code{[character(1)]} where to install packages.}
}
\description{
Install from GitHub style repositories
}
