% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-utilities.R
\name{pkg_github_ref}
\alias{pkg_github_ref}
\title{Get GitHub commit reference}
\usage{
pkg_github_ref(...)
}
\arguments{
\item{...}{\code{[character]} one or more package names.}
}
\value{
Returns a \code{[named:character]}. Names correspond to package names,
values correspond to GitHub SHA as determined from the \code{GithubRef} in the
package's DESCRIPTION file. Returns \code{NA} for packages without a \code{GithubRef}
and for uninstalled packages.
}
\description{
Vectorized GitHub commit reference determination for packages installed with
\code{devtools::install_github}, or packages with a \code{GithubRef} field in their
DESCRIPTION file.
}
