% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-utilities.R
\name{pkg_installed}
\alias{pkg_installed}
\title{Are packages installed?}
\usage{
pkg_installed(...)
}
\arguments{
\item{...}{\code{[character]} one or more package names.}
}
\value{
Returns a \code{[named:logical]}. Names correspond to package names,
values correspond to whether the package is installed (\code{TRUE/FALSE}).
Length of result corresponds to unique package names passed to \code{...}.
}
\description{
Vectorized check for package installation.
}
\examples{
\dontrun{
  others <- c('rmarkdown', 'ggplot2')
  is_installed <- .Rprofile$installed('tidyverse', 'devtools', others)
  is_installed['devtools']
}

}
