% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_dirs.R
\name{source_dirs}
\alias{source_dirs}
\title{Source all R scripts in a directory}
\usage{
source_dirs(dirs, lists = NULL)
}
\arguments{
\item{dirs}{\code{[character]} directories to source.}

\item{lists}{\code{[character]} names of empty list objects to create before
sourcing directories.}
}
\description{
Sources all R scripts in a directory in the order returned by \link{list.files}.
}
\details{
This function sources all scripts ending in \code{.R} or \code{.r} for each
directory specified in \code{dirs}. Empty list objects may be created with names
specified in \code{lists}. This allows scripts to assign to to these list
objects, which can be useful for modularizing functionality in a project.
}
